% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{Combine two factors}
\usage{
fbind(a, b)
}
\arguments{
\item{a}{A factor.}

\item{b}{A factor.}
}
\value{
A factor.
}
\description{
Coerces each input factor to character, combines them, then converts
back to a factor. New levels are the union of the levels of the inputs.
}
\examples{
a <- factor(c("character", "hits", "your", "eyeballs"))
b <- factor(c("but", "integer", "where it", "counts"))
fbind(a,b)
}
