% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_function.R
\name{hello_function}
\alias{hello_function}
\title{Say hello to my little function}
\usage{
hello_function(name)
}
\arguments{
\item{name}{character}
}
\value{
character
}
\description{
This function says hello.
}
\examples{
hello_function("Bond, James Bond")
}
