% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{Bind multiple factors together}
\usage{
fbind(...)
}
\arguments{
\item{...}{one or more \code{\link[base:factor]{factor}}}
}
\value{
a combined \code{factor()} that merges the unique levels in the \code{factor}
}
\description{
Bind multiple factors together
}
\examples{
#' #' ## Create 2 factors
a <- factor(c("character", "hits", "your", "eyeballs"))
b <- factor(c("but", "integer", "where it", "counts"))
c <- factor(c("always","count","your","eyeballs","too"))

## run fbind
fbind(a, b)
fbind(a, b, c)
}
