% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{Bind multiple factors}
\usage{
fbind(...)
}
\arguments{
\item{...}{factors}
}
\value{
combined factor
}
\description{
TODO
}
\examples{
x = c("1","2","3")
y = c("1","2","4")
v = c("12","24")

fbind(x,y,v)
}
